Change Attributes and AbilityValues to be doubles instead of ints. Any performance gains weren't worth the configuration pain.

+++ src/abilities/Ability.cs
@@ -214,7 +214,7 @@ public interface IAbilityCollection : IAbilityProvider
 public interface IAbilityContext : IAbilityCollection
 {
   public void GrantAbility(AbilityType ability);
-  public int GetNamedValue(string name);
+  public double GetNamedValue(string name);
 }
 
 public class ConcreteAbilityContext: IAbilityContext
@@ -237,7 +237,7 @@ public class ConcreteAbilityContext: IAbilityContext
     _abilities.Add(ability);
   }
 
-  public int GetNamedValue(string name)
+  public double GetNamedValue(string name)
   {
     return 0;
   }
